Move code to spec_test_components
I added a new spec_test_components file to not further increase the size of the spec_test_runner file. This PR removes other helper classes for the SpecTestRunner to this file. This was done using IntelliJ, I did not modify anything manually.
